Rename `<AssetDetailsDrawer>` to `<AssetDetailsPanel>`
