+++ src/components/AssetDetailsPanel/AssetDetailsPanel.vue
